pyproject.toml
[project]
name = "pyproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=2.24.0",
    "python-dotenv==1.1.0",
]
